working on upgrade priority unit test
