+++ server/test/constant/todo.go
@@ -1,0 +1,44 @@
+package constant_test
+
+import (
+	"bytes"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"io"
+	"net/http"
+	"net/http/httptest"
+	"sync"
+
+	"github.com/gin-gonic/gin"
+)
+
+func TodoCreate(wg *sync.WaitGroup, router *gin.Engine, requestBody []byte, cookie string) (interface{}, error) {
+	wg.Add(1)
+	recorder := httptest.NewRecorder()
+
+	go func() {
+		defer wg.Done()
+		request := httptest.NewRequest(http.MethodPost, "http://localhost:8080/api/v1/todo/", bytes.NewReader(requestBody))
+		request.Header.Add("Content-Type", "application/json")
+		request.Header.Add("Cookie", fmt.Sprintf("token=%v", cookie))
+		router.ServeHTTP(recorder, request)
+	}()
+
+	wg.Wait()
+
+	response := recorder.Result()
+	body, err := io.ReadAll(response.Body)
+	if err != nil {
+		return nil, err
+	}
+
+	var responseBody map[string]interface{}
+	json.Unmarshal(body, &responseBody)
+
+	if response.StatusCode == 200 {
+		return responseBody["data"], nil
+	}
+
+	return nil, errors.New(responseBody["data"].(string))
+}
